refactor: make enzyme setup opt in
