+++ .changeset/good-wolves-watch.md
@@ -1,0 +1,6 @@
+---
+'merchant-center-application-template-starter': patch
+'@commercetools-frontend/sdk': patch
+---
+
+Remove `enzyme` from tests and dependencies.
