+++ application-templates/starter/package.json
@@ -43,7 +43,6 @@
     "@manypkg/cli": "0.17.0",
     "@testing-library/react": "11.2.6",
     "dotenv-cli": "4.0.0",
-    "enzyme": "3.11.0",
     "jest": "26.6.3"
   }
 }
