+++ package.json
@@ -140,7 +140,6 @@
     "cross-env": "7.0.3",
     "cypress": "7.1.0",
     "dotenv": "8.2.0",
-    "enzyme": "3.11.0",
     "eslint": "7.25.0",
     "eslint-formatter-pretty": "4.0.0",
     "eslint-plugin-graphql": "4.0.0",
