+++ packages/jest-preset-mc-app/README.md
@@ -30,3 +30,85 @@ Supported options are:
 
 - `silenceConsoleWarnings`: a list of strings or RegEx to match against console logs. If the value matches, the log will be suppressed. This option only takes effect when the environment variable `CI` is defined.
 - `notThrowWarnings`: a list of strings or RegEx to match against console logs. If the value matches, no error will be thrown. This option only takes effect when the environment variable `CI` is defined.
+
+## TypeScript support
+
+If you need to use the Jest preset with additional TypeScript support, you need to import `@commercetools-frontend/jest-preset-mc-app/jest-preset-for-typescript.js` in your Jest config.
+
+For example:
+
+```js
+const jestPresetForTypeScript = require('@commercetools-frontend/jest-preset-mc-app/jest-preset-for-typescript');
+
+module.exports = {
+  preset: jestPresetForTypeScript,
+};
+```
+
+## Opt-in Enzyme support
+
+This Jest preset does not ship with Enzyme support anymore (version `>=20`).
+
+In case you still use Enzyme in your tests, you need to do a bit of extra configuration as described below:
+
+### Enzyme 16
+
+1. Install the necessary dependencies:
+
+```
+yarn add \
+  @commercetools/enzyme-extensions \
+  @commercetools/jest-enzyme-matchers \
+  enzyme \
+  enzyme-adapter-react-16 \
+  enzyme-matchers \
+  enzyme-to-json \
+  jest-enzyme
+```
+
+2. In your Jest config, instead of importing the preset `@commercetools-frontend/jest-preset-mc-app` you need to import `@commercetools-frontend/jest-preset-mc-app/enzyme/jest-preset-for-enzyme-16.js`
+
+For example:
+
+```js
+const jestPreset = require('@commercetools-frontend/jest-preset-mc-app');
+const jestPresetForEnzyme16 = require('@commercetools-frontend/jest-preset-mc-app/enzyme/jest-preset-for-enzyme-16');
+
+module.exports = {
+  preset: {
+    ...jestPreset,
+    ...jestPresetForEnzyme16,
+  },
+};
+```
+
+### Enzyme 17
+
+1. Install the necessary dependencies:
+
+```
+yarn add \
+  @commercetools/enzyme-extensions \
+  @commercetools/jest-enzyme-matchers \
+  @wojtekmaj/enzyme-adapter-react-17 \
+  enzyme \
+  enzyme-matchers \
+  enzyme-to-json \
+  jest-enzyme
+```
+
+2. In your Jest config, instead of importing the preset `@commercetools-frontend/jest-preset-mc-app` you need to import `@commercetools-frontend/jest-preset-mc-app/enzyme/jest-preset-for-enzyme-17.js`
+
+For example:
+
+```js
+const jestPreset = require('@commercetools-frontend/jest-preset-mc-app');
+const jestPresetForEnzyme17 = require('@commercetools-frontend/jest-preset-mc-app/enzyme/jest-preset-for-enzyme-17');
+
+module.exports = {
+  preset: {
+    ...jestPreset,
+    ...jestPresetForEnzyme17,
+  },
+};
+```
