+++ packages/jest-preset-mc-app/enzyme/create-jest-preset-for-enzyme.js
@@ -1,0 +1,16 @@
+const path = require('path');
+const defaultPreset = require('../jest-preset');
+
+const resolveRelativePath = (relativePath) =>
+  path.resolve(__dirname, relativePath);
+
+const createJestPresetForEnzyme = (version) => ({
+  ...defaultPreset,
+  snapshotSerializers: ['enzyme-to-json/serializer'],
+  setupFilesAfterEnv: [
+    resolveRelativePath('../setup-test-framework.js'),
+    resolveRelativePath(`./setup-test-framework-for-enzyme-${version}.js`),
+  ],
+});
+
+module.exports = createJestPresetForEnzyme;
