+++ packages/jest-preset-mc-app/enzyme/jest-preset-for-enzyme-17.js
@@ -1,0 +1,3 @@
+const createJestPresetForEnzyme = require('./create-jest-preset-for-enzyme');
+
+module.exports = createJestPresetForEnzyme(16);
