+++ packages/jest-preset-mc-app/enzyme/setup-test-framework-for-enzyme-16.js
@@ -1,0 +1,5 @@
+require('./base-setup-test-framework-for-enzyme');
+const Enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+
+Enzyme.configure({ adapter: new Adapter(), disableLifecycleMethods: true });
