+++ packages/jest-preset-mc-app/package.json
@@ -20,37 +20,24 @@
   "module": "./jest-preset.js",
   "dependencies": {
     "@commercetools-frontend/babel-preset-mc-app": "19.3.1",
-    "@commercetools/enzyme-extensions": "5.0.0",
-    "@commercetools/jest-enzyme-matchers": "1.1.2",
     "@sheerun/mutationobserver-shim": "0.3.3",
     "@testing-library/jest-dom": "5.12.0",
     "babel-jest": "26.6.3",
     "colors": "1.4.0",
     "cosmiconfig": "7.0.0",
-    "enzyme-adapter-react-16": "1.15.6",
-    "enzyme-matchers": "7.1.2",
-    "enzyme-to-json": "3.6.2",
     "identity-obj-proxy": "3.0.0",
     "intl": "1.2.5",
-    "jest-enzyme": "7.1.2",
     "jest-localstorage-mock": "2.4.10",
     "jest-silent-reporter": "0.5.0",
     "jest-transform-graphql": "2.1.0",
     "jest-watch-typeahead": "0.6.3",
     "make-plural": "6.2.2",
-    "pkg-dir": "5.0.0",
     "raf": "3.4.1",
     "ts-jest": "26.5.5",
     "unfetch": "4.2.0"
   },
   "devDependencies": {
     "@testing-library/react": "11.2.6",
-    "enzyme": "3.11.0",
     "jest": "26.6.3"
-  },
-  "peerDependencies": {
-    "@testing-library/react": "11.x",
-    "enzyme": "3.x",
-    "jest": "26.x"
   }
 }
