+++ packages/sdk/package.json
@@ -44,7 +44,6 @@
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@types/enzyme": "3.10.8",
     "@types/uuid": "8.3.0",
     "react": "17.0.2",
     "react-redux": "7.2.4",
